不要なvuefile を削除
